requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking
#
asgiref==3.7.2
    # via django
dj-database-url==2.0.0
    # via -r requirements.in
django==4.2.2
    # via
    #   -r zion/requirements.in
    #   dj-database-url
    #   django-appconf
    #   django-extensions
django-appconf==1.0.5
    # via -r zion/requirements.in
django-extensions==3.2.3
    # via -r zion/requirements.in
django-split-settings==1.2.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
packaging==23.1
    # via gunicorn
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.6.3
    # via dj-database-url
